requirements.txt
Django==2.0.1
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Markdown==2.6.11
mysqlclient==1.3.12
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
virtualenv==15.1.0
